Pointer changes 3

+++ printf/ft_hex.c
@@ -6,7 +6,7 @@
 /*   By: rprieto- <rprieto-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/12 22:52:14 by rprieto-          #+#    #+#             */
-/*   Updated: 2020/10/14 21:03:39 by rprieto-         ###   ########.fr       */
+/*   Updated: 2020/10/14 21:04:56 by rprieto-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -100,7 +100,7 @@ void	print_pointer(void *pointer, t_modifiers modifiers, int *char_sum)
 {
 	int pointer_length;
 
-	pointer_length = get_hex_digits(pointer);
+	pointer_length = get_hex_digits((long)pointer);
 	*char_sum += write(1, "0x", 2);
 	if (pointer == NULL)
 	{
